move analytics to top level
